fix: adjust dex sort rule (#5434)
* fix: adjust sort rule

* fix: bugfix

* chore: code style

* fix: set auto complete to off

* Update packages/mask/src/plugins/Trader/SNSAdaptor/trader/hooks/useGasConfig.ts


* Update packages/mask/src/plugins/Trader/SNSAdaptor/trader/hooks/useGasConfig.ts
