+++ packages/mask/src/plugins/Trader/SNSAdaptor/trader/InputTokenPanel.tsx
@@ -184,7 +184,7 @@ export const InputTokenPanel = memo<InputTokenPanelProps>(
                         </Typography>
                     ),
                 }}
-                inputProps={{ className: classes.input }}
+                inputProps={{ className: classes.input, autoComplete: 'off' }}
             />
         )
     },
