+++ packages/mask/src/plugins/Trader/SNSAdaptor/trader/Trader.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useEffect, useMemo, useState } from 'react'
+import { useCallback, useEffect, useState } from 'react'
 import { makeStyles, useStylesExtends } from '@masknet/theme'
 import {
     ChainId,
@@ -6,46 +6,37 @@ import {
     createNativeToken,
     EthereumTokenType,
     formatBalance,
-    formatGweiToWei,
     FungibleTokenDetailed,
-    GasOptionConfig,
-    getNetworkTypeFromChainId,
     isSameAddress,
     TransactionStateType,
     useChainId,
     useChainIdValid,
     useFungibleTokenBalance,
-    useProviderType,
     useTokenConstants,
     useWallet,
-    useAccount,
 } from '@masknet/web3-shared-evm'
-import { isGreaterThan, isLessThan, multipliedBy } from '@masknet/web3-shared-base'
 import { useRemoteControlledDialog } from '@masknet/shared'
 import { delay } from '@masknet/shared-base'
+import { useGasConfig } from './hooks/useGasConfig'
 import type { Coin } from '../../types'
 import { TokenPanelType, TradeInfo } from '../../types'
 import { useI18N } from '../../../../utils'
 import { TradeForm } from './TradeForm'
 import { AllProviderTradeActionType, AllProviderTradeContext } from '../../trader/useAllProviderTradeContext'
-import { MINIMUM_AMOUNT, UST } from '../../constants'
+import { UST } from '../../constants'
 import { SelectTokenDialogEvent, WalletMessages } from '@masknet/plugin-wallet'
-import { useAsync, useUnmount, useUpdateEffect } from 'react-use'
+import { useUnmount, useUpdateEffect } from 'react-use'
 import { isTwitter } from '../../../../social-network-adaptor/twitter.com/base'
 import { activatedSocialNetworkUI } from '../../../../social-network'
 import { isFacebook } from '../../../../social-network-adaptor/facebook.com/base'
 import { useTradeCallback } from '../../trader/useTradeCallback'
 import { isNativeTokenWrapper } from '../../helpers'
 import { ConfirmDialog } from './ConfirmDialog'
-import Services from '../../../../extension/service'
-import { currentBalancesSettings } from '../../../Wallet/settings'
 import { TargetChainIdContext } from '../../trader/useTargetChainIdContext'
-import { WalletRPC } from '../../../Wallet/messages'
 import { PluginTraderMessages } from '../../messages'
-import { NetworkType } from '@masknet/public-api'
-import BigNumber from 'bignumber.js'
-import { useNativeTokenPrice, useTokenPrice } from '../../../Wallet/hooks/useTokenPrice'
 import { SettingsDialog } from './SettingsDialog'
+import { useSortedTrades } from './hooks/useSortedTrades'
+import { useUpdateBalance } from './hooks/useUpdateBalance'
 
 const useStyles = makeStyles()(() => {
     return {
@@ -65,14 +56,11 @@ export function Trader(props: TraderProps) {
     const { coin, tokenDetailed, chainId: targetChainId } = props
     const { decimals } = tokenDetailed ?? coin ?? {}
     const [focusedTrade, setFocusTrade] = useState<TradeInfo>()
-    const [gasConfig, setGasConfig] = useState<GasOptionConfig | undefined>()
     const wallet = useWallet()
     const currentChainId = useChainId()
     const chainId = targetChainId ?? currentChainId
     const chainIdValid = useChainIdValid()
     const { NATIVE_TOKEN_ADDRESS } = useTokenConstants()
-    const currentAccount = useAccount()
-    const currentProvider = useProviderType()
     const classes = useStylesExtends(useStyles(), props)
     const { t } = useI18N()
     const { setTargetChainId } = TargetChainIdContext.useContainer()
@@ -85,7 +73,11 @@ export function Trader(props: TraderProps) {
         ],
         allTradeComputed,
     } = AllProviderTradeContext.useContainer()
-    //endregion
+    //#endregion
+
+    //#region gas config and gas price
+    const { gasPrice, gasConfig, setGasConfig } = useGasConfig(chainId)
+    //#endregion
 
     //#region if chain id be changed, update input token be native token
     useEffect(() => {
@@ -177,56 +169,6 @@ export function Trader(props: TraderProps) {
         loadingOutputTokenBalance,
         NATIVE_TOKEN_ADDRESS,
     ])
-
-    // Query the balance of native tokens on target chain
-    useAsync(async () => {
-        if (!currentAccount) {
-            dispatchTradeStore({
-                type: AllProviderTradeActionType.UPDATE_INPUT_TOKEN_BALANCE,
-                balance: '0',
-            })
-
-            dispatchTradeStore({
-                type: AllProviderTradeActionType.UPDATE_OUTPUT_TOKEN_BALANCE,
-                balance: '0',
-            })
-            return
-        }
-
-        if (chainId && currentProvider && currentAccount) {
-            const cacheBalance = currentBalancesSettings.value[currentProvider]?.[chainId]
-
-            let balance: string
-
-            if (cacheBalance) balance = cacheBalance
-            else {
-                balance = await Services.Ethereum.getBalance(currentAccount, {
-                    chainId: chainId,
-                    providerType: currentProvider,
-                })
-                await WalletRPC.updateBalances({
-                    [currentProvider]: {
-                        [chainId]: balance,
-                    },
-                })
-            }
-
-            dispatchTradeStore({
-                type: AllProviderTradeActionType.UPDATE_INPUT_TOKEN_BALANCE,
-                balance: inputToken?.type === EthereumTokenType.Native ? balance : '0',
-            })
-
-            dispatchTradeStore({
-                type: AllProviderTradeActionType.UPDATE_OUTPUT_TOKEN_BALANCE,
-                balance:
-                    isSameAddress(outputToken?.address, NATIVE_TOKEN_ADDRESS) ||
-                    outputToken?.type === EthereumTokenType.Native
-                        ? balance
-                        : '0',
-            })
-        }
-    }, [inputToken, outputToken, currentAccount, currentProvider, chainId, currentChainId])
-    // #endregion
 
     // #region select token
     const excludeTokens = [inputToken, outputToken].filter(Boolean).map((x) => x?.address) as string[]
@@ -354,43 +296,6 @@ export function Trader(props: TraderProps) {
     )
     //#endregion
 
-    //#region get gas price
-    const { value: gasPrice } = useAsync(async () => {
-        try {
-            const network = getNetworkTypeFromChainId(chainId)
-            if (gasConfig) {
-                return new BigNumber(
-                    (network === NetworkType.Ethereum ? gasConfig.maxFeePerGas : gasConfig.gasPrice) ?? 0,
-                ).toString()
-            } else {
-                if (network === NetworkType.Ethereum) {
-                    const response = await WalletRPC.getEstimateGasFees(chainId)
-                    const maxFeePerGas = formatGweiToWei(response?.medium?.suggestedMaxFeePerGas ?? 0).toString()
-                    const maxPriorityFeePerGas = formatGweiToWei(
-                        response?.medium?.suggestedMaxPriorityFeePerGas ?? 0,
-                    ).toString()
-                    setGasConfig({
-                        maxFeePerGas,
-                        maxPriorityFeePerGas,
-                    })
-
-                    return maxFeePerGas
-                } else {
-                    const response = await WalletRPC.getGasPriceDictFromDeBank(chainId)
-                    const gasPrice = new BigNumber(response?.data.normal.price ?? 0).toString()
-                    setGasConfig({
-                        gasPrice,
-                    })
-
-                    return gasPrice
-                }
-            }
-        } catch {
-            return '0'
-        }
-    }, [chainId, gasConfig])
-    //#endregion
-
     //#region open the transaction dialog
     useEffect(() => {
         if (tradeState?.type === TransactionStateType.UNKNOWN) return
@@ -411,53 +316,12 @@ export function Trader(props: TraderProps) {
     //#endregion
 
     //#region The trades sort by best price (Estimate received * price - Gas fee * native token price)
-    const nativeTokenPrice = useNativeTokenPrice(chainId)
-    const outputTokenPrice = useTokenPrice(chainId, outputToken?.address.toLowerCase())
-    const sortedAllTradeComputed = useMemo(() => {
-        if (outputToken && (outputTokenPrice || nativeTokenPrice)) {
-            return allTradeComputed
-                .map((trade) => {
-                    if (
-                        gasPrice &&
-                        trade.value &&
-                        isGreaterThan(trade.value.outputAmount, MINIMUM_AMOUNT) &&
-                        trade.gas.value
-                    ) {
-                        const gasFee = multipliedBy(gasPrice, trade.gas.value).integerValue().toFixed()
-
-                        const gasFeeUSD = new BigNumber(formatBalance(gasFee ?? 0, outputToken.decimals)).times(
-                            nativeTokenPrice,
-                        )
-
-                        const finalPrice = new BigNumber(
-                            formatBalance(trade.value.outputAmount, outputToken.decimals, 2),
-                        )
-                            .times(outputToken.type !== EthereumTokenType.Native ? outputTokenPrice : nativeTokenPrice)
-                            .minus(gasFeeUSD)
-
-                        return {
-                            ...trade,
-                            finalPrice,
-                        }
-                    }
-                    return trade
-                })
-                .filter(({ value }) => !!value && !value.outputAmount.isZero())
-                .sort(({ finalPrice: a }, { finalPrice: b }) => {
-                    if (a && b && isGreaterThan(a, b)) return -1
-                    if (a && b && isLessThan(a, b)) return 1
-                    return 0
-                })
-        }
-        return allTradeComputed
-            .filter(({ value }) => !!value && !value.outputAmount.isZero())
-            .sort(({ value: a }, { value: b }) => {
-                if (a?.outputAmount.isGreaterThan(b?.outputAmount ?? 0)) return -1
-                if (a?.outputAmount.isLessThan(b?.outputAmount ?? 0)) return 1
-                return 0
-            })
-    }, [allTradeComputed, outputToken, gasPrice, outputTokenPrice, nativeTokenPrice])
+    const sortedAllTradeComputed = useSortedTrades(allTradeComputed, chainId, gasPrice)
     //#endregion
+
+    // Query the balance of native tokens on target chain
+    useUpdateBalance(chainId, currentChainId)
+    // #endregion
 
     //#region reset focused trade when chainId, inputToken, outputToken, inputAmount be changed
     useUpdateEffect(() => {
